refactor(PopularStudy): 스터디 데이터 커스텀 훅으로 추상화
